feat: created impure pipe
impure pipes can cause performance issues. However in order to update
the object (which pipe displays) you need to either recreate the object
on each modification (may be prefered way to go about it) or add pure:
false for the pipe. Under the hood impure pipe get an event trigger on
which they update